Use getopt_long for option parsing
boost::program_options started to look way to heavy for a simple
stand alone utility like this, so although it's a bit more work to
use getopt, it makes the build a lot easier.

Remove the couple of places where a C++14 construct was
used, to drop the required C++ version to 11. This allows building
with compilers as old as gcc4.8 (e.g., native on CentOS7).

Removed doxygen build elements from the CMakeLists files as
this project does not need them.

Update README.