Improve CPU time monitoring unittest (#14)
* Improve CPU time monitoring unittest

Write CPU monitor test in python as it's far easier to
check the results are sane. Add 3 tests to the CMakefile
that test single process, multi-process and multi-thead
running.

* Improve test program

Support proper passing of all relvant arguments, with appropriate defaults
In case of failure print expectations

* Handle bad arguments properly

It was possible to set negative values for threads/procs. This was particularly
bad with an unsigned variable, giving a value >> 1.