Fix bug in parsing smaps files and add memory tests
smaps file was being processed without resetting to the
beginning of the line when trying to get key:value pairs.
Now after the k:v read the rest of the line is ignored.

Add a mem-burner program that allocates memory, writes
to only some of it and forks some child processes. This can test
prmon's measurements of vmem, rss and pss. Add two
unit tests for single process and multi process cases.