Reimplement old PID parser for older kernels
Old kernels do not have /proc exposure of their child PIDs
so on these systems it's necessary to fall back to the previously
used "pstree" parsing (this is less efficient and more fragile,
but it does work well enough; https://lwn.net/Articles/475688/).

To avoid testing on every iteration, a single test for the kernel's
support of the new interface is done once, then a bool stored
that switches between the implementations for old/new kernels
in the monitoring loop.

There was enough different PID parsing code now to justify its
own source file.