Better way to build statically
This is still a WIP, because it can be improved in its implementation,
but let me illustrate the idea. In this case we support a special CMake
flag that enables a static build by adding the correct linker flag for
prmon iff the flag is set via set_target_properties().

This prevents a global change that affects the test binaries and it
sits orthogonal to the main CMAKE_BUILD_TYPE, which is actually
quite useful as debug symbols and optimisation flags get controlled
independently.

This basic idea can be extended to also support flags for gprof and
gperftools, which is quite nice.