Remove multi-threaded handling for SIGUSR1
The handling of SIGUSR1 was using multiple threads and
a condition variable to achieve an almost instant exit when
prmon recieved a signal of this type. Although the code was
correct, it was causing some issues with the static build in
Ubuntu 16.04 (core dumps).

As the benefit of this type of exit strategy was rather marginal
(at most speeding up by 1s) and the vast majority of exits
are actually from the monitored process finishing (which is
only tested in the polling loop itself), it did not seem worthwhile
keeping this. Instead, tighten the polling loop up to 250ms.

However, we keep the same behaviour - on SIGUSR1 exit
writing the current JSON summary file.

(Note also that that that multi-thread strategy was not at all
free - it required spawning a thread on each loop, which
is relatively expensive.)