Add a CMake build option to generate a static binary and support profiling (#63)
* Add a CMake build option to generate a static binary

For a utility like prmon having a statically build binary is
a very useful thing, so add the correct compiler and linker
options to do that.

* Better way to build statically

This is still a WIP, because it can be improved in its implementation,
but let me illustrate the idea. In this case we support a special CMake
flag that enables a static build by adding the correct linker flag for
prmon iff the flag is set via set_target_properties().

This prevents a global change that affects the test binaries and it
sits orthogonal to the main CMAKE_BUILD_TYPE, which is actually
quite useful as debug symbols and optimisation flags get controlled
independently.

This basic idea can be extended to also support flags for gprof and
gperftools, which is quite nice.

* Support static and profiled builds for prmon

Add three specific build variables that allow prmon to be
build as a static binary, or using gprof or gperftools profiling
enabled.

* Remove multi-threaded handling for SIGUSR1

The handling of SIGUSR1 was using multiple threads and
a condition variable to achieve an almost instant exit when
prmon recieved a signal of this type. Although the code was
correct, it was causing some issues with the static build in
Ubuntu 16.04 (core dumps).

As the benefit of this type of exit strategy was rather marginal
(at most speeding up by 1s) and the vast majority of exits
are actually from the monitored process finishing (which is
only tested in the polling loop itself), it did not seem worthwhile
keeping this. Instead, tighten the polling loop up to 250ms.

However, we keep the same behaviour - on SIGUSR1 exit
writing the current JSON summary file.

(Note also that that that multi-thread strategy was not at all
free - it required spawning a thread on each loop, which
is relatively expensive.)

* Add a CMake find package for GPerfTools

As this is not a standard "find" for any CMake install I tested
I added a fairly generic FindGperftools CMake script. This was
tweaked slightly from

https://raw.githubusercontent.com/apache/incubator-mxnet/master/cmake/Modules/FindGperftools.cmake

to allow linking to just the profile libray (no tcmalloc).

Gperftools_ROOT_DIR can be used to hint to CMake when
the gperftools are in an odd place.