Handle correctly situation where /etc/os-release is incomplete
If the parsing of /etc/os-release failed to set either HSF_OS_ID or
HSF_OS_VERSION there was then a nasty CMake error thrown
in a later REGEX REPLACE. At least for bleeding edge releases
the VERSION_ID string can be absent (noticed on Gentoo portage)
so this broke.

In these cases we set the strings to some value that's consistent with
not knowing the correct value ("unknown v0").