Implement average values as doubles
Having average value calculations as unsigned integers was producing harsh
rounding errors for small valued quantities (like nprocs, nthreads). These are
now implemented as double precision values and get serialised as
numbers in JSON with decimal fractions.

There is a slight drawback that the decimal part is designed in nlohmann::json
to avoid precision loss, so sometimes a very high number of DPs are printed.
However, there doesn't seem to be any way to adjust this without manually
massaging the actual numerical values (there are various issues in
https://github.com/nlohmann/json where this is discussed).