Add NVIDIA GPU monitoring (#125)
* Add GPU monitoring for NVIDIA cards

cmd_pipe_output():
- Use low level pipe+fork+dup2, for execution of external process
- Calling signature to vector of strings for caller (cmd + args), which is a lot
 easier; conversion to suitable char* values is handled internally
test_nvidia_smi():
- Parse output sensibly to see if there are GPUs to monitor and print a warning
 if there are more than 4 (which renders 'pmon' unreliable)

Add a new virtual is_valid() function to the Imonitor interface,
with simple 'return true;' for most monitors, but using the
GPU check code for the nvidia monitor

Add NVIDIA_DEBUG #define for debugging any issues

Add GPU hardware information. As the GPUs in the system could, in
principle, be different we store
each one separately, but also store the total number. At the moment
we take the model name, the SM clock frequency and the memory
as useful parameters.

* clang-format fixes

* Reduce code complexity measure

Reduce 'if' statements to jump out of loops when possible
Use a map instead of a vector to make gpu counting less complex
Ensure that all read variables are initialised so that they have
some value in them in case of a read abort

* Small fixes after review

Convert tabs to spaces
Add error reporting if polling fails
Fix logic in parsing output from nvidia-smi pmon
Use a descriptive constant instead of a magic number for MB->KB