Implement fix for missing i/o values
Under some circumstances that are not very well
understood the process tree chain loses i/o values
(which would normally always be "inherited" by
parent processes as children exit).

To compensate for this the iomon module will now
record maximum i/o stats, and if the polled values
are below these it will replace the stats with the
previous maximums. A warning is also printed.

To test this logic there is a defined variable in
iomon.cpp that switches on code that artificially
provokes the problem, exercising the recovery
logic.